% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewp.R
\name{dewp2}
\alias{dewp2}
\title{Probability mass function of the two-parameter EWP}
\usage{
dewp2(x, lambda, beta)
}
\arguments{
\item{x}{vector of (positive integer) quantiles.}

\item{lambda}{centrality parameter}

\item{beta}{dispersion parameter}
}
\value{

}
\description{
Probability mass function of the two-parameter EWP
}
