% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewp_reg.R
\name{ewp_reg}
\alias{ewp_reg}
\title{Exponentially weighted poisson regression model}
\usage{
ewp_reg(
  formula,
  family = "ewp3",
  data,
  verbose = TRUE,
  method = "BFGS",
  hessian = TRUE,
  autoscale = TRUE,
  maxiter = 500
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{family}{choice of "ewp2" or "ewp3"}

\item{data}{a data frame containing the variables in the model.}

\item{verbose}{logical, defaults to TRUE; print model fitting progress}

\item{hessian}{logical, defaults to TRUE; calculate Hessian?}

\item{autoscale}{logical, defaults to TRUE; automatically scale model parameters inside the optimisation routine based on initial estimates from a Poisson regression.}

\item{maxiter}{numeric maximum number of iterations for optim}
}
\value{

}
\description{
Exponentially weighted poisson regression model
}
