% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewp_reg.R
\name{mmean}
\alias{mmean}
\title{Estimate marginal means}
\usage{
mmean(object, cov, ...)
}
\arguments{
\item{object}{ewp model object}

\item{cov}{character; covariate to find marginal mean for}

\item{...}{ignored}
}
\value{
printout of the marginal means
}
\description{
Estimate marginal means
}
