% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewp_reg.R
\name{summary.ewp}
\alias{summary.ewp}
\title{Model summary}
\usage{
\method{summary}{ewp}(object, ...)
}
\arguments{
\item{object}{ewp model fit}

\item{...}{ignored}
}
\value{
The function `summary.ewp` computes and returns a list of summary statistics of the fitted ewp model.
}
\description{
Model summary
}
